Practical/Shopify/requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asn1crypto==1.5.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
Babel==2.10.3
backcall==0.2.0
bcrypt==3.2.2
beautifulsoup4==4.12.2
billiard==3.6.4.0
bleach==6.0.0
blinker==1.6.2
Brlapi==0.8.4
Brotli==1.0.9
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==5.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloud-init==23.1.2
colorama==0.4.6
comm==0.1.3
command-not-found==0.3
configobj==5.0.8
cron-descriptor==1.3.0
cryptography==38.0.4
cssselect2==0.7.0
cupshelpers==1.0
dbus-python==1.3.2
debugpy==1.6.7
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.8.0
distro-info==1.5
Django==4.2.1
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-timezone-field==5.0
djangorestframework==3.14.0
duplicity==0.8.22
executing==1.2.0
fasteners==0.17.3
fastjsonschema==2.16.3
Flask==2.3.2
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-MySQL==1.5.2
Flask-MySQLdb==1.0.1
flask-paginate==2022.1.8
Flask-RESTful==0.3.9
fonttools==4.39.4
fqdn==1.5.1
future==0.18.2
html5lib==1.1
httplib2==0.20.4
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
isoduration==20.11.0
itsdangerous==2.1.2
jaraco.classes==3.2.1
jedi==0.18.2
jeepney==0.8.0
Jinja2==3.1.2
json5==0.9.11
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab_server==2.22.1
keyring==23.9.3
kombu==5.2.4
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.5
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.24.0
lxml==4.9.2
Mako==1.2.4.dev0
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==2.0.5
monotonic==1.6
more-itertools==8.10.0
mysql-connector==2.2.9
mysql-connector-python==8.0.33
mysqlclient==2.1.1
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
netifaces==0.11.0
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
olefile==0.46
oscrypto==1.3.0
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
paramiko==2.12.0
parso==0.8.3
pdfkit==1.0.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.5.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
pydyf==0.6.0
Pygments==2.14.0
PyGObject==3.44.1
pyHanko==0.18.1
pyhanko-certvalidator==0.22.0
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pypdf==3.8.1
pyphen==0.14.0
pypng==0.20220715.0
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.5.3+ubuntu1
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-json-logger==2.0.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyxdg==0.28
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
redis==4.5.5
reportlab==3.6.12
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
SecretStorage==3.3.3
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
svglib==1.5.1
systemd-python==235
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
uri-template==1.2.0
uritools==4.0.1
urllib3==1.26.12
usb-creator==0.3.16
vine==5.0.0
wadllib==1.3.6
wcwidth==0.2.6
weasyprint==59.0
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.3.3
xdg==5
xhtml2pdf==0.2.11
xkit==0.0.0
y-py==0.5.9
ypy-websocket==0.8.2
zipp==1.0.0
zopfli==0.2.2
